+++ crates/client/src/app/input/mod.rs
@@ -1,3 +1,4 @@
+pub mod body;
 pub mod crafting;
 pub mod hotbar_assignment;
 pub mod inventory;
